Relocate app class in separate file and refactor main()
